requirements.txt
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
python-decouple==3.8

